Move to immediate mode parsing API
This enables the parsing to be tested easily
