+++ imrichtext.cpp
@@ -1,17 +1,17 @@
 #include "ImRichText.h"
+#if __has_include("imrichtextfont.h")
+#define IM_RICHTEXT_DEFAULT_FONTS_AVAILABLE
 #include "imrichtextfont.h"
-#include "imrichtextcolor.h"
+#endif
 #include "imgui_internal.h"
-
+#include "imrichtextutils.h"
 #include <unordered_map>
 #include <cstring>
 #include <cctype>
 #include <optional>
 #include <cstdlib>
 #include <string>
-#include <tuple>
 #include <chrono>
-#include <array>
 
 #ifdef _DEBUG
 #include <cstdio>
@@ -49,7 +49,6 @@ namespace ImRichText
     struct BlockquoteDrawData
     {
         std::vector<std::pair<ImVec2, ImVec2>> bounds;
-        float linewidth = 0.f;
     };
 
     struct RichTextData
@@ -80,13 +79,26 @@ namespace ImRichText
         std::pair<int, int> end;
     };
 
+    struct StackData
+    {
+        std::string_view tag;
+        int styleIdx = -1;
+    };
+
+    struct BackgroundData
+    {
+        BackgroundSpanData span;
+        int styleIdx = -1;
+    };
+
     static std::unordered_map<std::string_view, std::size_t> RichTextStrings;
     static std::unordered_map<std::size_t, RichTextData> RichTextMap;
-    static std::pair<std::string_view, SegmentStyle> TagStack[IM_RICHTEXT_MAXDEPTH];
-    static std::vector<BackgroundSpanData> BgLineStartIndex[IM_RICHTEXT_MAXDEPTH];
     static std::unordered_map<std::string_view, AnimationData> AnimationMap;
 
+    static StackData TagStack[IM_RICHTEXT_MAXDEPTH];
+    static std::vector<BackgroundData> BackgroundSpans[IM_RICHTEXT_MAXDEPTH];
     static int CurrentStackPos = -1;
+
     static int ListItemCountByDepths[IM_RICHTEXT_MAX_LISTDEPTH];
     static BlockquoteDrawData BlockquoteStack[IM_RICHTEXT_MAXDEPTH];
 
@@ -105,48 +117,6 @@ namespace ImRichText
 
     static const char* LineSpaces = "                                ";
 
-    template <typename T>
-    [[nodiscard]] T Clamp(T val, T min, T max)
-    {
-        return val < min ? min : val > max ? max : val;
-    }
-
-#pragma optimize( "", on )
-    [[nodiscard]] int SkipSpace(const char* text, int idx, int end)
-    {
-        while ((idx < end) && std::isspace(text[idx])) idx++;
-        return idx;
-    }
-
-    [[nodiscard]] int SkipSpace(const std::string_view text, int from = 0)
-    {
-        auto end = (int)text.size();
-        while ((from < end) && (std::isspace(text[from]))) from++;
-        return from;
-    }
-
-    [[nodiscard]] int WholeWord(const std::string_view text, int from = 0)
-    {
-        auto end = (int)text.size();
-        while ((from < end) && (!std::isspace(text[from]))) from++;
-        return from;
-    }
-
-    [[nodiscard]] int SkipDigits(const std::string_view text, int from = 0)
-    {
-        auto end = (int)text.size();
-        while ((from < end) && (std::isdigit(text[from]))) from++;
-        return from;
-    }
-
-    [[nodiscard]] int SkipFDigits(const std::string_view text, int from = 0)
-    {
-        auto end = (int)text.size();
-        while ((from < end) && ((std::isdigit(text[from])) || (text[from] == '.'))) from++;
-        return from;
-    }
-#pragma optimize( "", off )
-
 #ifndef IMGUI_DEFINE_MATH_OPERATORS
     [[nodiscard]] ImVec2 operator*(ImVec2 lhs, float rhs)
     {
@@ -164,48 +134,6 @@ namespace ImRichText
     }
 #endif
 
-    [[nodiscard]] bool AreSame(const std::string_view lhs, const char* rhs)
-    {
-        auto rlimit = (int)std::strlen(rhs);
-        auto llimit = (int)lhs.size();
-        if (rlimit != llimit)
-            return false;
-
-        for (auto idx = 0; idx < rlimit; ++idx)
-            if (std::tolower(lhs[idx]) != std::tolower(rhs[idx]))
-                return false;
-
-        return true;
-    }
-
-    [[nodiscard]] bool StartsWith(const std::string_view lhs, const char* rhs)
-    {
-        auto rlimit = (int)std::strlen(rhs);
-        auto llimit = (int)lhs.size();
-        if (rlimit > llimit)
-            return false;
-
-        for (auto idx = 0; idx < rlimit; ++idx)
-            if (std::tolower(lhs[idx]) != std::tolower(rhs[idx]))
-                return false;
-
-        return true;
-    }
-
-    [[nodiscard]] bool AreSame(const std::string_view lhs, const std::string_view rhs)
-    {
-        auto rlimit = (int)rhs.size();
-        auto llimit = (int)lhs.size();
-        if (rlimit != llimit)
-            return false;
-
-        for (auto idx = 0; idx < rlimit; ++idx)
-            if (std::tolower(lhs[idx]) != std::tolower(rhs[idx]))
-                return false;
-
-        return true;
-    }
-
 #ifdef _DEBUG
     [[nodiscard]] const char* GetTokenTypeString(const Token& token)
     {
@@ -220,283 +148,14 @@ namespace ImRichText
     }
 #endif
 
-    [[nodiscard]] int ExtractInt(std::string_view input, int defaultVal)
+    template <typename T>
+    [[nodiscard]] T Clamp(T val, T min, T max)
     {
-        int result = defaultVal;
-        int base = 1;
-        auto idx = (int)input.size() - 1;
-        while (!std::isdigit(input[idx]) && idx >= 0) idx--;
-
-        for (; idx >= 0; --idx)
-        {
-            result += (input[idx] - '0') * base;
-            base *= 10;
-        }
-
-        return result;
+        return val < min ? min : val > max ? max : val;
     }
 
-    [[nodiscard]] int ExtractIntFromHex(std::string_view input, int defaultVal)
-    {
-        int result = defaultVal;
-        int base = 1;
-        auto idx = (int)input.size() - 1;
-        while (!std::isdigit(input[idx]) && !std::isalpha(input[idx]) && idx >= 0) idx--;
-
-        for (; idx >= 0; --idx)
-        {
-            result += std::isdigit(input[idx]) ? (input[idx] - '0') * base : 
-                std::islower(input[idx]) ? ((input[idx] - 'a') + 10) * base : 
-                ((input[idx] - 'A') + 10) * base;
-            base *= 16;
-        }
-
-        return result;
-    }
-
-    struct IntOrFloat
-    {
-        float value = 0.f;
-        bool isFloat = false;
-    };
-
-    [[nodiscard]] IntOrFloat ExtractNumber(std::string_view input, float defaultVal)
-    {
-        float result = 0.f, base = 1.f;
-        bool isInt = false;
-        auto idx = (int)input.size() - 1;
-
-        while (idx >= 0 && input[idx] != '.') idx--;
-        auto decimal = idx;
-
-        if (decimal != -1)
-        {
-            for (auto midx = decimal; midx >= 0; --midx)
-            {
-                result += (input[midx] - '0') * base;
-                base *= 10.f;
-            }
-
-            base = 0.1f;
-            for (auto midx = decimal + 1; midx < (int)input.size(); ++midx)
-            {
-                result += (input[midx] - '0') * base;
-                base *= 0.1f;
-            }
-        }
-        else
-        {
-            for (auto midx = (int)input.size() - 1; midx >= 0; --midx)
-            {
-                result += (input[midx] - '0') * base;
-                base *= 10.f;
-            }
-
-            isInt = true;
-        }
-
-        return { result, !isInt };
-    }
-
-    [[nodiscard]] float ExtractFloatWithUnit(std::string_view input, float defaultVal, float ems, float parent, float scale)
-    {
-        float result = defaultVal, base = 1.f;
-        auto idx = (int)input.size() - 1;
-        while (!std::isdigit(input[idx]) && idx >= 0) idx--;
-
-        if (AreSame(input.substr(idx + 1), "pt")) scale = 1.3333f;
-        else if (AreSame(input.substr(idx + 1), "em")) scale = ems;
-        else if (input[idx + 1] == '%') scale = parent * 0.01f;
-
-        auto num = ExtractNumber(input.substr(0, idx + 1), defaultVal);
-        result = num.value;
-
-        return result * scale;
-    }
-
-    [[nodiscard]] std::tuple<IntOrFloat, IntOrFloat, IntOrFloat, IntOrFloat> GetCommaSeparatedNumbers(std::string_view stylePropVal, int& curr)
-    {
-        std::tuple<IntOrFloat, IntOrFloat, IntOrFloat, IntOrFloat> res;
-        auto hasFourth = curr == 4;
-        curr = SkipSpace(stylePropVal, curr);
-        assert(stylePropVal[curr] == '(');
-        curr++;
-
-        auto valstart = curr;
-        curr = SkipFDigits(stylePropVal, curr);
-        std::get<0>(res) = ExtractNumber(stylePropVal.substr(valstart, curr - valstart), 0);
-        curr = SkipSpace(stylePropVal, curr);
-        assert(stylePropVal[curr] == ',');
-        curr++;
-        curr = SkipSpace(stylePropVal, curr);
-
-        valstart = curr;
-        curr = SkipFDigits(stylePropVal, curr);
-        std::get<1>(res) = ExtractNumber(stylePropVal.substr(valstart, curr - valstart), 0);
-        curr = SkipSpace(stylePropVal, curr);
-        assert(stylePropVal[curr] == ',');
-        curr++;
-        curr = SkipSpace(stylePropVal, curr);
-
-        valstart = curr;
-        curr = SkipFDigits(stylePropVal, curr);
-        std::get<2>(res) = ExtractNumber(stylePropVal.substr(valstart, curr - valstart), 0);
-        curr = SkipSpace(stylePropVal, curr);
-
-        if (hasFourth)
-        {
-            assert(stylePropVal[curr] == ',');
-            curr++;
-            curr = SkipSpace(stylePropVal, curr);
-
-            valstart = curr;
-            curr = SkipFDigits(stylePropVal, curr);
-            std::get<3>(res) = ExtractNumber(stylePropVal.substr(valstart, curr - valstart), 0);
-        }
-
-        return res;
-    }
-
-    [[nodiscard]] ImColor ExtractColor(std::string_view stylePropVal, ImColor(*NamedColor)(const char*, void*), void* userData)
-    {
-        if (stylePropVal.size() >= 3u && AreSame(stylePropVal.substr(0, 3), "rgb"))
-        {
-            IntOrFloat r, g, b, a;
-            auto hasAlpha = stylePropVal[3] == 'a' || stylePropVal[3] == 'A';
-            int curr = hasAlpha ? 4 : 3;
-            std::tie(r, g, b, a) = GetCommaSeparatedNumbers(stylePropVal, curr);
-            auto isRelative = r.isFloat && g.isFloat && b.isFloat;
-            a.value = isRelative ? hasAlpha ? a.value : 1.f :
-                hasAlpha ? a.value : 255;
-
-            assert(stylePropVal[curr] == ')');
-            return isRelative ? ImColor{ r.value, g.value, b.value, a.value } : 
-                ImColor{ (int)r.value, (int)g.value, (int)b.value, (int)a.value };
-        }
-        else if (stylePropVal.size() >= 3u && AreSame(stylePropVal.substr(0, 3), "hsv"))
-        {
-            IntOrFloat h, s, v;
-            auto curr = 3;
-            std::tie(h, s, v, std::ignore) = GetCommaSeparatedNumbers(stylePropVal, curr);
-            
-            assert(stylePropVal[curr] == ')');
-            return ImColor::HSV(h.value, s.value, v.value);
-        }
-        else if (stylePropVal.size() >= 3u && AreSame(stylePropVal.substr(0, 3), "hsl"))
-        {
-            IntOrFloat h, s, l;
-            auto curr = 3;
-            std::tie(h, s, l, std::ignore) = GetCommaSeparatedNumbers(stylePropVal, curr);
-            auto v = l.value + s.value * std::min(l.value, 1.f - l.value);
-            s.value = v == 0.f ? 0.f : 2.f * (1.f - (l.value / v));
-
-            assert(stylePropVal[curr] == ')');
-            return ImColor::HSV(h.value, s.value, v);
-        }
-        else if (stylePropVal.size() >= 1u && stylePropVal[0] == '#')
-        {
-            ImU32 color32 = (ImU32)ExtractIntFromHex(stylePropVal.substr(1), 0);
-            return ImColor{ color32 };
-        }
-        else if (NamedColor != nullptr)
-        {
-            static char buffer[32] = { 0 };
-            std::memset(buffer, 0, 32);
-            std::memcpy(buffer, stylePropVal.data(), std::min((int)stylePropVal.size(), 31));
-            return NamedColor(buffer, userData);
-        }
-        else
-        {
-            return IM_COL32_BLACK;
-        }
-    }
-
-    std::pair<ImColor, float> ExtractColorStop(std::string_view input, const RenderConfig& config)
-    {
-        auto idx = 0;
-        std::pair<ImColor, float> colorstop;
-        
-        idx = WholeWord(input, idx);
-        colorstop.first = ExtractColor(input.substr(0, idx), config.NamedColor, config.UserData);
-        idx = SkipSpace(input, idx);
-
-        if ((idx < (int)input.size()) && std::isdigit(input[idx]))
-        {
-            auto start = idx;
-            idx = SkipDigits(input, start);
-            colorstop.second = ExtractNumber(input.substr(start, idx - start), -1.f).value;
-        }
-        else colorstop.second = -1.f;
-
-        return colorstop;
-    }
-
-    ColorGradient ExtractLinearGradient(std::string_view input, const RenderConfig& config)
-    {
-        ColorGradient gradient;
-        auto idx = 15; // size of "linear-gradient" string
-
-        if (idx < (int)input.size())
-        {
-            idx = SkipSpace(input, idx);
-            assert(input[idx] == '(');
-            idx++;
-            idx = SkipSpace(input, idx);
-            
-            std::optional<std::pair<ImColor, float>> lastStop = std::nullopt;
-            auto firstPart = true;
-            auto start = idx;
-            auto total = 0.f, unspecified = 0.f;
-
-            do
-            {
-                idx = SkipSpace(input, idx);
-
-                auto start = idx;
-                while ((idx < (int)input.size()) && (input[idx] != ',') && (input[idx] != ')') 
-                    && !std::isspace(input[idx])) idx++;
-                auto part = input.substr(start, idx - start);
-                idx = SkipSpace(input, idx);
-                auto isEnd = input[idx] == ')';
-
-                if ((idx < (int)input.size()) && (input[idx] == ',' || isEnd)) {
-                    if (firstPart)
-                    {
-                        if (AreSame(input, "to right")) gradient.reverseOrder = true;
-                        else if (!AreSame(input, "to left")) {
-                            auto colorstop = ExtractColorStop(part, config);
-                            if (colorstop.second != -1.f) total += colorstop.second;
-                            else unspecified += 1.f;
-                            lastStop = colorstop;
-                        }
-                        firstPart = false;
-                    }
-                    else {
-                        auto colorstop = ExtractColorStop(part, config);
-                        if (colorstop.second != -1.f) total += colorstop.second;
-                        else unspecified += 1.f;
-                        if (lastStop.has_value()) gradient.colorStops.emplace_back(
-                            ColorStop{ lastStop.value().first, colorstop.first, colorstop.second });
-                        lastStop = colorstop;
-                    }
-                }
-                else break;
-                
-                if (isEnd) break;
-                else if (input[idx] == ',') idx++;
-            } while ((idx < (int)input.size()) && (input[idx] != ')'));
-
-            unspecified -= 1.f;
-            for (auto& colorstop : gradient.colorStops)
-                if (colorstop.pos == -1.f) colorstop.pos = (100.f - total) / (100.f * unspecified);
-                else colorstop.pos /= 100.f;
-        }
-
-        return gradient;
-    }
-
-    int PopulateSegmentStyle(SegmentStyle& style,
-        const SegmentStyle& parentStyle,
+    int PopulateSegmentStyle(StyleDescriptor& style,
+        const StyleDescriptor& parentStyle,
         std::string_view stylePropName,
         std::string_view stylePropVal,
         const RenderConfig& config)
@@ -542,7 +201,7 @@ namespace ImRichText
         else if (AreSame(stylePropName, "background-color") || AreSame(stylePropName, "background"))
         {
             if (StartsWith(stylePropVal, "linear-gradient")) 
-               style.gradient = ExtractLinearGradient(stylePropVal, config);
+               style.gradient = ExtractLinearGradient(stylePropVal, config.NamedColor, config.UserData);
             else style.bgcolor = ExtractColor(stylePropVal, config.NamedColor, config.UserData);
             prop = StyleBackground;
         }
@@ -671,61 +330,11 @@ namespace ImRichText
         return sz;
     }
 
-    [[nodiscard]] std::optional<std::string_view> GetQuotedString(const char* text, int& idx, int end)
+    [[nodiscard]] StyleDescriptor CreateDefaultStyle(const RenderConfig& config)
     {
-        auto insideQuotedString = false;
-        auto begin = idx;
+        assert(config.GetFont != nullptr);       
 
-        if ((idx < end) && text[idx] == '\'')
-        {
-            idx++;
-
-            while (idx < end)
-            {
-                if (text[idx] == '\\' && ((idx + 1 < end) && text[idx + 1] == '\''))
-                {
-                    insideQuotedString = !insideQuotedString;
-                    idx++;
-                }
-                else if (!insideQuotedString && text[idx] == '\'') break;
-                idx++;
-            }
-        }
-        else if ((idx < end) && text[idx] == '"')
-        {
-            idx++;
-
-            while (idx < end)
-            {
-                if (text[idx] == '\\' && ((idx + 1 < end) && text[idx + 1] == '"'))
-                {
-                    insideQuotedString = !insideQuotedString;
-                    idx++;
-                }
-                else if (!insideQuotedString && text[idx] == '"') break;
-                idx++;
-            }
-        }
-
-        if ((idx < end) && (text[idx] == '"' || text[idx] == '\''))
-        {
-            std::string_view res{ text + begin + 1, (std::size_t)(idx - begin - 1) };
-            idx++;
-            return res;
-        }
-        else if (idx > begin)
-        {
-            ERROR("Quoted string invalid... [%.*s] at %d\n", (int)(idx - begin), text + begin, idx);
-        }
-
-        return std::nullopt;
-    }
-
-    [[nodiscard]] SegmentStyle CreateDefaultStyle(const RenderConfig& config)
-    {
-        assert(config.GetFont != nullptr);
-
-        SegmentStyle result;
+        StyleDescriptor result;
         result.font.family = config.DefaultFontFamily;
         result.font.size = config.DefaultFontSize * config.FontScale;
         result.font.font = config.GetFont(result.font.family, result.font.size, false, false, false, config.UserData);
@@ -735,9 +344,12 @@ namespace ImRichText
         return result;
     }
 
-    [[nodiscard]] ImVec2 GetSegmentSize(const SegmentData& segment, const RenderConfig& config)
+    [[nodiscard]] ImVec2 GetSegmentSize(const SegmentData& segment, const std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         auto height = 0.f, width = 0.f;
+        if (segment.StyleIdx == -1) return { 0.f, 0.f };
+
+        const auto& style = styles[segment.StyleIdx + 1];
 
         for (const auto& token : segment.Tokens)
         {
@@ -745,45 +357,44 @@ namespace ImRichText
             width += token.Bounds.width;
         }
 
-        return { width, height };
+        return { width + style.padding.h(), height + style.padding.v() };
     }
 
-    [[nodiscard]] ImVec2 GetLineSize(const DrawableLine& line, const RenderConfig& config)
+    [[nodiscard]] ImVec2 GetLineSize(const DrawableLine& line, const std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         auto height = 0.f, width = 0.f;
 
         for (const auto& segment : line.Segments)
         {
-            auto sz = GetSegmentSize(segment, config);
-            auto offseth = segment.Style.subscriptOffset + segment.Style.superscriptOffset +
-                segment.Style.padding.top + segment.Style.padding.bottom;
-            height = std::max(height, sz.y + offseth);
-            width += sz.x + segment.Style.padding.left + segment.Style.padding.right;
+            auto sz = GetSegmentSize(segment, styles, config);
+            height = std::max(height, sz.y);
+            width += sz.x;
         }
 
         return { width, height };
     }
 
-    void AddToken(DrawableLine& line, Token token, int propsChanged, const RenderConfig& config)
+    void AddToken(DrawableLine& line, Token token, int propsChanged, const std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         auto& segment = line.Segments.back();
+        const auto& style = styles[segment.StyleIdx + 1];
 
         if (token.Type == TokenType::Text)
         {
-            auto sz = config.GetTextSize(token.Content, segment.Style.font.font);
+            auto sz = config.GetTextSize(token.Content, style.font.font);
             token.Bounds.width = sz.x;
             token.Bounds.height = sz.y;
         }
         else if (token.Type == TokenType::HorizontalRule)
         {
-            token.Bounds.width = config.Bounds.x - line.Content.left - (line.Offset.right + line.Offset.left) - 
-                (segment.Style.padding.left + segment.Style.padding.right);
-            token.Bounds.height = segment.Style.height;
+            token.Bounds.width = config.Bounds.x - line.Content.left - line.Offset.h() -
+                style.padding.h();
+            token.Bounds.height = style.height;
         }
         else if (token.Type == TokenType::ListItemBullet)
         {
             auto bulletscale = Clamp(config.BulletSizeScale, 1.f, 4.f);
-            auto bulletsz = (segment.Style.font.size) / bulletscale;
+            auto bulletsz = (style.font.size) / bulletscale;
             token.Bounds.width = token.Bounds.height = bulletsz;
         }
         else if (token.Type == TokenType::ListItemNumbered)
@@ -811,7 +422,7 @@ namespace ImRichText
             }
 
             std::string_view input{ token.NestedListItemIndex, (size_t)currbuf };
-            auto sz = config.GetTextSize(input, segment.Style.font.font);
+            auto sz = config.GetTextSize(input, style.font.font);
             token.Bounds.width = sz.x;
             token.Bounds.height = sz.y;
         }
@@ -831,30 +442,30 @@ namespace ImRichText
 
         LOG("Added token: %.*s [itemtype: %s][font-size: %f][size: (%f, %f)]\n",
             (int)token.Content.size(), token.Content.data(),
-            GetTokenTypeString(token), segment.Style.font.size,
+            GetTokenTypeString(token), style.font.size,
             token.Bounds.width, token.Bounds.height);
     }
 
-    SegmentData& AddSegment(DrawableLine& line, const SegmentStyle& styleprops, const RenderConfig& config)
+    SegmentData& AddSegment(DrawableLine& line, int styleIdx, const std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         if (!line.Segments.empty())
         {
             auto& lastSegment = line.Segments.back();
-            auto sz = GetSegmentSize(lastSegment, config);
+            auto sz = GetSegmentSize(lastSegment, styles, config);
             lastSegment.Bounds.width = sz.x;
             lastSegment.Bounds.height = sz.y;
         }
 
         auto& segment = line.Segments.emplace_back();
-        segment.Style = styleprops;
+        segment.StyleIdx = styleIdx;
         return segment;
     }
 
-    [[nodiscard]] DrawableLine CreateNewLine(const SegmentStyle& styleprops, const RenderConfig& config)
+    [[nodiscard]] DrawableLine CreateNewLine(int styleIdx)
     {
         DrawableLine line;
         line.Segments.emplace_back();
-        line.Segments.back().Style = styleprops;
+        line.Segments.back().StyleIdx = styleIdx;
         line.BlockquoteDepth = -1;
         return line;
     }
@@ -889,7 +500,7 @@ namespace ImRichText
     };
 
     std::vector<int> PerformWordWrap(std::deque<DrawableLine>& lines, TagType tagType, int index, int listDepth, int blockquoteDepth,
-        const RenderConfig& config)
+        const std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         std::vector<int> result;
         if (!lines[index].HasText || !config.WordWrap || !config.GetTextSize)
@@ -899,27 +510,29 @@ namespace ImRichText
         }
 
         std::deque<DrawableLine> newlines;
-        auto currline = CreateNewLine(lines.front().Segments.front().Style, config);
+        auto currline = CreateNewLine(-1);
         auto currentx = 0.f;
         auto availwidth = lines[index].Content.width;
 
         for (const auto& segment : lines[index].Segments)
         {
+            const auto& style = styles[segment.StyleIdx + 1];
+
             for (const auto& token : segment.Tokens)
             {
                 if (token.Type == TokenType::Text)
                 {
-                    auto sz = config.GetTextSize(token.Content, segment.Style.font.font);
+                    auto sz = config.GetTextSize(token.Content, style.font.font);
 
                     if (currentx + sz.x > availwidth)
                     {
                         newlines.push_back(currline);
                         currline.Segments.clear();
-                        currline.Segments.emplace_back();
+                        currline.Segments.emplace_back().StyleIdx = segment.StyleIdx;
                         currentx = 0.f;
                     }
 
-                    currline.Segments.back().Style = segment.Style;
+                    currline.Segments.back().StyleIdx = segment.StyleIdx;
                     currline.Segments.back().Tokens.emplace_back(token);
                     currentx += sz.x;
                 }
@@ -944,7 +557,7 @@ namespace ImRichText
         return sum * (baseFontSz * 0.5f);
     }
 
-    float GetMaxSuperscriptOffset(const DrawableLine& line, float scale)
+    float GetMaxSuperscriptOffset(const DrawableLine& line, const std::vector<StyleDescriptor>& styles, float scale)
     {
         auto topOffset = 0.f;
         auto baseFontSz = 0.f;
@@ -952,7 +565,7 @@ namespace ImRichText
         for (auto idx = 0; idx < (int)line.Segments.size();)
         {
             const auto& segment = line.Segments[idx];
-            baseFontSz = segment.Style.font.size;
+            baseFontSz = styles[segment.StyleIdx + 1].font.size;
             auto depth = 0, begin = idx;
 
             while ((idx < (int)line.Segments.size()) && (line.Segments[idx].SuperscriptDepth > 0))
@@ -968,7 +581,7 @@ namespace ImRichText
         return topOffset;
     }
 
-    float GetMaxSubscriptOffset(const DrawableLine& line, float scale)
+    float GetMaxSubscriptOffset(const DrawableLine& line, const std::vector<StyleDescriptor>& styles, float scale)
     {
         auto topOffset = 0.f;
         auto baseFontSz = 0.f;
@@ -976,7 +589,7 @@ namespace ImRichText
         for (auto idx = 0; idx < (int)line.Segments.size();)
         {
             const auto& segment = line.Segments[idx];
-            baseFontSz = segment.Style.font.size;
+            baseFontSz = styles[segment.StyleIdx + 1].font.size;
             auto depth = 0, begin = idx;
 
             while ((idx < (int)line.Segments.size()) && (line.Segments[idx].SubscriptDepth > 0))
@@ -992,49 +605,52 @@ namespace ImRichText
         return topOffset;
     }
 
-    void AdjustForSuperSubscripts(const std::vector<int>& indexes, std::deque<DrawableLine>& lines, const RenderConfig& config)
+    void AdjustForSuperSubscripts(const std::vector<int>& indexes, std::deque<DrawableLine>& lines, 
+        std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         for (auto idx : indexes)
         {
             auto& line = lines[idx];
             if (!line.HasSubscript && !line.HasSuperscript) continue;
 
-            auto maxTopOffset = GetMaxSuperscriptOffset(line, config.ScaleSuperscript);
-            auto maxBottomOffset = GetMaxSubscriptOffset(line, config.ScaleSubscript);
+            auto maxTopOffset = GetMaxSuperscriptOffset(line, styles, config.ScaleSuperscript);
+            auto maxBottomOffset = GetMaxSubscriptOffset(line, styles, config.ScaleSubscript);
             auto lastFontSz = config.DefaultFontSize * config.FontScale;
             auto lastSuperscriptDepth = 0, lastSubscriptDepth = 0;
 
             for (auto& segment : line.Segments)
             {
+                auto& style = styles[segment.StyleIdx + 1];
+
                 if (segment.SuperscriptDepth > lastSuperscriptDepth)
                 {
-                    segment.Style.font.size = lastFontSz * config.ScaleSuperscript;
-                    maxTopOffset -= segment.Style.font.size * 0.5f;
+                    style.font.size = lastFontSz * config.ScaleSuperscript;
+                    maxTopOffset -= style.font.size * 0.5f;
                 }
                 else if (segment.SuperscriptDepth < lastSuperscriptDepth)
                 {
                     maxTopOffset += lastFontSz * 0.5f;
-                    segment.Style.font.size = lastFontSz / config.ScaleSuperscript;
+                    style.font.size = lastFontSz / config.ScaleSuperscript;
                 }
 
                 if (segment.SubscriptDepth > lastSubscriptDepth)
                 {
-                    segment.Style.font.size = lastFontSz * config.ScaleSubscript;
-                    maxBottomOffset += (lastFontSz - segment.Style.font.size * 0.5f);
+                    style.font.size = lastFontSz * config.ScaleSubscript;
+                    maxBottomOffset += (lastFontSz - style.font.size * 0.5f);
                 }
                 else if (segment.SubscriptDepth < lastSubscriptDepth)
                 {
-                    segment.Style.font.size = lastFontSz / config.ScaleSubscript;
-                    maxBottomOffset -= segment.Style.font.size * 0.5f;
+                    style.font.size = lastFontSz / config.ScaleSubscript;
+                    maxBottomOffset -= style.font.size * 0.5f;
                 }
 
-                segment.Style.superscriptOffset = maxTopOffset;
-                segment.Style.subscriptOffset = maxBottomOffset;
+                style.superscriptOffset = maxTopOffset;
+                style.subscriptOffset = maxBottomOffset;
                 segment.Bounds.height += maxTopOffset + maxBottomOffset;
 
                 lastSuperscriptDepth = segment.SuperscriptDepth;
                 lastSubscriptDepth = segment.SubscriptDepth;
-                lastFontSz = segment.Style.font.size;
+                lastFontSz = style.font.size;
             }
         }
     }
@@ -1050,13 +666,13 @@ namespace ImRichText
     }
 
     void ComputeLineBounds(std::deque<DrawableLine>& result, const std::vector<int>& linesModified, 
-        const RenderConfig& config)
+        const std::vector<StyleDescriptor>& styles, const RenderConfig& config)
     {
         for (auto index : linesModified)
         {
             auto& line = result[index];
             auto currx = line.Content.left + line.Offset.left;
-            auto sz = GetLineSize(line, config);
+            auto sz = GetLineSize(line, styles, config);
             line.Content.width = sz.x;
             line.Content.height = sz.y;
 
@@ -1064,17 +680,20 @@ namespace ImRichText
 
             for (auto& segment : line.Segments)
             {
+                if (segment.Tokens.empty()) continue;
+
                 // This will align the segment as vertical centre, TODO: Handle other alignments
                 segment.Bounds.top = line.Content.top + line.Offset.top + ((line.Content.height - segment.height()) * 0.5f);
                 segment.Bounds.left = currx; // This will align left, TODO: Handle other alignments
+                const auto& style = styles[segment.StyleIdx + 1];
 
-                currx += segment.Style.padding.left;
+                currx += style.padding.left;
 
                 for (auto tokidx = 0; tokidx < (int)segment.Tokens.size(); ++tokidx)
                 {
                     auto& token = segment.Tokens[tokidx];
-                    token.Bounds.top = segment.Bounds.top + segment.Style.padding.top +
-                        segment.Style.superscriptOffset + segment.Style.subscriptOffset +
+                    token.Bounds.top = segment.Bounds.top + style.padding.top +
+                        style.superscriptOffset + style.subscriptOffset +
                         ((segment.Bounds.height - token.Bounds.height) * 0.5f);
 
                     // TODO: Fix bullet positioning w.r.t. first text block (baseline aligned?)
@@ -1084,7 +703,7 @@ namespace ImRichText
                     currx += token.Bounds.width;
                 }
 
-                currx += segment.Style.padding.right;
+                currx += style.padding.right;
             }
 
             HIGHLIGHT("\nCreated line #%d at (%f, %f) of size (%f, %f) with %d segments", index,
@@ -1093,8 +712,9 @@ namespace ImRichText
         }
     }
 
-    DrawableLine MoveToNextLine(const SegmentStyle& styleprops, TagType tagType, int listDepth, int blockquoteDepth,
-        DrawableLine& line, std::deque<DrawableLine>& result, const RenderConfig& config, bool isTagStart)
+    DrawableLine MoveToNextLine(TagType tagType, int listDepth, int blockquoteDepth, int styleIdx,
+        DrawableLine& line, std::deque<DrawableLine>& result, std::vector<StyleDescriptor>& styles,
+        const RenderConfig& config, bool isTagStart)
     {
         auto isEmpty = IsLineEmpty(line);
         result.emplace_back(line);
@@ -1108,13 +728,13 @@ namespace ImRichText
         else
         {
             if (!line.Marquee && config.Bounds.x > 0.f)
-                linesModified = PerformWordWrap(result, tagType, (int)result.size() - 1, listDepth, blockquoteDepth, config);
+                linesModified = PerformWordWrap(result, tagType, (int)result.size() - 1, listDepth, blockquoteDepth, styles, config);
             else linesModified.push_back((int)result.size() - 1);
-            AdjustForSuperSubscripts(linesModified, result, config);
+            AdjustForSuperSubscripts(linesModified, result, styles, config);
         }
 
         auto& lastline = result.back();
-        auto newline = CreateNewLine(styleprops, config);
+        auto newline = CreateNewLine(styleIdx);
         newline.BlockquoteDepth = blockquoteDepth;
         if (isTagStart) newline.Marquee = tagType == TagType::Marquee;
 
@@ -1122,7 +742,7 @@ namespace ImRichText
         if (blockquoteDepth > lastline.BlockquoteDepth) newline.Offset.top = config.BlockquotePadding;
         else if (blockquoteDepth < lastline.BlockquoteDepth) lastline.Offset.bottom = config.BlockquotePadding;
 
-        ComputeLineBounds(result, linesModified, config);
+        ComputeLineBounds(result, linesModified, styles, config);
 
         newline.Content.left = ((float)(listDepth + 1) * config.ListItemIndent) + 
             ((float)(blockquoteDepth + 1) * config.BlockquoteOffset);
@@ -1131,11 +751,12 @@ namespace ImRichText
     }
 
     void GenerateTextToken(DrawableLine& line, std::string_view content,
-        int curridx, int start, const RenderConfig& config)
+        int curridx, int start, const std::vector<StyleDescriptor>& styles, 
+        const RenderConfig& config)
     {
         Token token;
         token.Content = content.substr(start, curridx - start);
-        AddToken(line, token, NoStyleChange, config);
+        AddToken(line, token, NoStyleChange, styles, config);
     }
 
     [[nodiscard]] std::pair<bool, bool> AddEscapeSequences(const std::string_view content,
@@ -1187,7 +808,7 @@ namespace ImRichText
     }
 
     int RecordTagProperties(TagType tagType, std::string_view attribName, std::optional<std::string_view> attribValue, 
-        SegmentStyle& style, const SegmentStyle& parentStyle, const RenderConfig& config)
+        StyleDescriptor& style, const StyleDescriptor& parentStyle, const RenderConfig& config)
     {
         int result = 0;
 
@@ -1243,107 +864,28 @@ namespace ImRichText
             if (AreSame(attribName, "min") && attribValue.has_value()) style.range.first = ExtractInt(attribValue.value(), 0);
             if (AreSame(attribName, "max") && attribValue.has_value()) style.range.second = ExtractInt(attribValue.value(), 0);
         }
-    }
 
-    int ExtractAttributes(const char* text, int end, std::string_view currTag, TagType tagType, const RenderConfig& config, SegmentStyle& style, const SegmentStyle& initialStyle, int& idx)
-    {
-        const auto& parentStyle = CurrentStackPos > 0 ? TagStack[CurrentStackPos - 1].second : initialStyle;
-        int result = NoStyleChange;
-
-        // Extract all attributes
-        while ((idx < end) && (text[idx] != config.TagEnd) && (text[idx] != '/'))
-        {
-            auto begin = idx;
-            while ((idx < end) && (text[idx] != '=') && !std::isspace(text[idx]) && (text[idx] != '/')) idx++;
-
-            if (text[idx] != '/')
-            {
-                auto attribName = std::string_view{ text + begin, (std::size_t)(idx - begin) };
-                LOG("Attribute: %.*s\n", (int)attribName.size(), attribName.data());
-
-                idx = SkipSpace(text, idx, end);
-                if (text[idx] == '=') idx++;
-                idx = SkipSpace(text, idx, end);
-                auto attribValue = GetQuotedString(text, idx, end);
-
-                
-            }
-        }
-
-        if (text[idx] == config.TagEnd) idx++;
-        if (text[idx] == '/' && ((idx + 1) < end) && text[idx + 1] == config.TagEnd) idx += 2;
         return result;
     }
 
-    std::pair<std::string_view, bool> ExtractTag(const char* text, int end, char TagEnd,
-        int& idx, bool& tagStart)
+    int CreateNextStyle(std::vector<StyleDescriptor>& styles)
     {
-        std::pair<std::string_view, bool> result;
-        result.second = true;
-
-        if (text[idx] == '/')
-        {
-            tagStart = false;
-            idx++;
-        }
-        else if (!std::isalnum(text[idx]))
-        {
-            ERROR("Invalid tag at %d...\n", idx);
-            result.second = false;
-            return result;
-        }
-
-        auto begin = idx;
-        while ((idx < end) && !std::isspace(text[idx]) && (text[idx] != TagEnd)) idx++;
-
-        if (idx - begin == 0)
-        {
-            ERROR("Empty tag at %d...\n", begin);
-            result.second = false;
-            return result;
-        }
-
-        result.first = std::string_view{ text + begin, (std::size_t)(idx - begin) };
-        if (result.first.back() == '/') result.first = result.first.substr(0, result.first.size() - 1u);
-
-        if (!tagStart)
-        {
-            if (text[idx] == TagEnd) idx++;
-
-            if (result.first.empty())
-            {
-                ERROR("Empty tag at %d...\n", begin);
-                result.second = false;
-                return result;
-            }
-            if (CurrentStackPos > -1 && result.first != TagStack[CurrentStackPos].first)
-            {
-                ERROR("Closing tag <%.*s> doesnt match opening tag <%.*s>...",
-                    (int)result.first.size(), result.first.data(), (int)TagStack[CurrentStackPos].first.size(),
-                    TagStack[CurrentStackPos].first.data());
-                result.second = false;
-                return result;
-            }
-        }
-
-        idx = SkipSpace(text, idx, end);
-        return result;
+        auto& newstyle = styles.emplace_back(styles.back());
+        newstyle.bgcolor = IM_COL32_BLACK_TRANS;
+        newstyle.gradient = ColorGradient{};
+        return (int)styles.size() - 1;
     }
 
-    SegmentStyle& PushTag(std::string_view currTag, const SegmentStyle& initialStyle)
+    void PushTag(std::string_view currTag, int styleIdx)
     {
         CurrentStackPos++;
-        TagStack[CurrentStackPos].first = currTag;
-        TagStack[CurrentStackPos].second = CurrentStackPos > 0 ?
-            TagStack[CurrentStackPos - 1].second : initialStyle;
-        TagStack[CurrentStackPos].second.gradient.colorStops.clear();
-        TagStack[CurrentStackPos].second.bgcolor = IM_COL32_BLACK_TRANS;
-        return TagStack[CurrentStackPos].second;
+        TagStack[CurrentStackPos].tag = currTag;
+        TagStack[CurrentStackPos].styleIdx = styleIdx;
     }
 
     void PopTag(bool reset)
     {
-        if (reset) TagStack[CurrentStackPos] = std::make_pair(std::string_view{}, SegmentStyle{});
+        if (reset) TagStack[CurrentStackPos].tag = "";
         --CurrentStackPos;
     }
 
@@ -1378,7 +920,7 @@ namespace ImRichText
     }
 
     void SetImplicitStyleProps(TagType tagType, std::string_view currTag, 
-        SegmentStyle& styleprops, const SegmentStyle& parentStyle,
+        StyleDescriptor& styleprops, const StyleDescriptor& parentStyle,
         DrawableLine& line, int propsChanged, const RenderConfig& config)
     {
         assert(config.GetFont != nullptr);
@@ -1451,7 +993,11 @@ namespace ImRichText
             propsChanged = propsChanged | StyleFontStyle;
             propsChanged = propsChanged | StyleFgColor;
         }
-        else if (tagType == TagType::Blink) styleprops.blink = true;
+        else if (tagType == TagType::Blink)
+        {
+            styleprops.blink = true;
+            propsChanged = propsChanged | StyleBlink;
+        }
         else if (tagType == TagType::Meter)
         {
             if ((propsChanged & StyleBackground) == 0) styleprops.bgcolor = config.MeterBgColor;
@@ -1478,16 +1024,55 @@ namespace ImRichText
 #endif
 
     template <typename ItrT>
-    void DrawLinearGradient(ImDrawList* drawList, ImVec2 startpos, ImVec2 endpos, ItrT start, ItrT end)
+    void DrawLinearGradient(ImDrawList* drawList, ImVec2 initpos, ImVec2 endpos, float angle, ImGuiDir dir, ItrT start, ItrT end)
     {
-        auto width = endpos.x - startpos.x;
+        auto width = endpos.x - initpos.x;
+        auto height = endpos.y - initpos.y;
 
-        for (auto it = start; it != end; ++it)
+        if (dir == ImGuiDir::ImGuiDir_Left)
         {
-            auto extent = width * it->pos;
-            drawList->AddRectFilledMultiColor(startpos, startpos + ImVec2{ extent, (endpos.y - startpos.y) },
-                it->from, it->to, it->to, it->from);
-            startpos.x += extent;
+            if (angle != 0.f)
+            {
+                ImVec2 points[4];
+                ImU32 colors[4];
+
+                for (auto it = start; it != end; ++it)
+                {
+                    auto extent = width * it->pos;
+                    auto m = std::tanf(angle);
+
+                    points[0] = initpos;
+                    points[1] = points[0] + ImVec2{ extent, 0.f };
+                    points[2] = initpos - ImVec2{ m * initpos.y, height };
+                    points[3] = points[2] + ImVec2{ extent, 0.f };
+
+                    colors[0] = colors[3] = it->from;
+                    colors[1] = colors[2] = it->to;
+
+                    DrawPolyFilledMultiColor(drawList, points, colors, 4);
+                    initpos.x += extent;
+                }
+            }
+            else
+            {
+                for (auto it = start; it != end; ++it)
+                {
+                    auto extent = width * it->pos;
+                    drawList->AddRectFilledMultiColor(initpos, initpos + ImVec2{ extent, height },
+                        it->from, it->to, it->to, it->from);
+                    initpos.x += extent;
+                }
+            }
+        }
+        else if (dir == ImGuiDir::ImGuiDir_Down)
+        {
+            for (auto it = start; it != end; ++it)
+            {
+                auto extent = height * it->pos;
+                drawList->AddRectFilledMultiColor(initpos, initpos + ImVec2{ width, extent },
+                    it->from, it->from, it->to, it->to);
+                initpos.y += extent;
+            }
         }
     }
 
@@ -1496,16 +1081,16 @@ namespace ImRichText
     {
         if (!gradient.colorStops.empty())
             (!gradient.reverseOrder) ?
-            DrawLinearGradient(drawList, startpos, endpos,
+            DrawLinearGradient(drawList, startpos, endpos, gradient.angleDegrees, gradient.dir,
                 gradient.colorStops.begin(), gradient.colorStops.end()) :
-            DrawLinearGradient(drawList, startpos, endpos,
+            DrawLinearGradient(drawList, startpos, endpos, gradient.angleDegrees, gradient.dir,
                 gradient.colorStops.rbegin(), gradient.colorStops.rend());
         else if (color.Value != config.DefaultBgColor.Value && color.Value != ImColor{ IM_COL32_BLACK_TRANS }.Value)
             drawList->AddRectFilled(startpos, endpos, color);
     }
 
     bool DrawToken(ImDrawList* drawList, int lineidx, const Token& token, ImVec2 initpos,
-        ImVec2 bounds, const SegmentStyle& style, const RenderConfig& config, 
+        ImVec2 bounds, const StyleDescriptor& style, const RenderConfig& config, 
         TooltipData& tooltip, AnimationData& animation, bool isMarquee)
     {
         if (style.blink && !animation.isVisible) return true;
@@ -1520,103 +1105,15 @@ namespace ImRichText
             auto bulletscale = Clamp(config.BulletSizeScale, 1.f, 4.f);
             auto bulletsz = (style.font.size) / bulletscale;
 
-            switch (style.list.itemStyle)
+            if (style.list.itemStyle == BulletType::Custom)
             {
-            case BulletType::Circle: {
-                ImVec2 center = token.Bounds.center(initpos);
-                drawList->AddCircle(center, bulletsz * 0.5f, style.fgcolor);
-                break;
-            }
-
-            case BulletType::Disk: {
-                ImVec2 center = token.Bounds.center(initpos);
-                drawList->AddCircleFilled(center, bulletsz * 0.5f, style.fgcolor);
-                break;
-            }
-
-            case BulletType::Square: {
-                drawList->AddRectFilled(token.Bounds.start(initpos), token.Bounds.end(initpos), style.fgcolor);
-                break;
-            }
-
-            case BulletType::Concentric: {
-                ImVec2 center = token.Bounds.center(initpos);
-                drawList->AddCircle(center, bulletsz * 0.5f, style.fgcolor);
-                drawList->AddCircleFilled(center, bulletsz * 0.4f, style.fgcolor);
-                break;
-            }
-
-            case BulletType::Triangle: {
-                auto startpos = token.Bounds.start(initpos);
-                auto offset = bulletsz * 0.25f;
-                ImVec2 a{ startpos.x, startpos.y },
-                       b{ startpos.x + bulletsz, startpos.y + (bulletsz * 0.5f) },
-                       c{ startpos.x, startpos.y + bulletsz };
-                drawList->AddTriangleFilled(a, b, c, style.fgcolor);
-                break;
-            }
-
-            case BulletType::Arrow: {
-                auto startpos = token.Bounds.start(initpos);
-                auto bsz2 = bulletsz * 0.5f;
-                auto bsz3 = bulletsz * 0.33333f;
-                auto bsz6 = bsz3 * 0.5f;
-                auto bsz38 = bulletsz * 0.375f;
-                ImVec2 points[7];
-                points[0] = { startpos.x, startpos.y + bsz38 };
-                points[1] = { startpos.x + bsz2, startpos.y + bsz38 };
-                points[2] = { startpos.x + bsz2, startpos.y + bsz6 };
-                points[3] = { startpos.x + bulletsz, startpos.y + bsz2 };
-                points[4] = { startpos.x + bsz2, startpos.y + bulletsz - bsz6 };
-                points[5] = { startpos.x + bsz2, startpos.y + bulletsz - bsz38 };
-                points[6] = { startpos.x, startpos.y + bulletsz - bsz38 };
-                drawList->AddRectFilled(points[0], points[5], style.fgcolor);
-                drawList->AddTriangleFilled(points[2], points[3], points[4], style.fgcolor);
-                break;
-            }
-
-            case BulletType::CheckMark: {
-                auto startpos = token.Bounds.start(initpos);
-                auto bsz3 = (bulletsz * 0.25f);
-                auto thickness = bulletsz * 0.2f;
-                ImVec2 points[3];
-                points[0] = { startpos.x, startpos.y + (2.5f * bsz3) };
-                points[1] = { startpos.x + (bulletsz * 0.3333f), startpos.y + bulletsz};
-                points[2] = { startpos.x + bulletsz, startpos.y + bsz3 };
-                drawList->AddPolyline(points, 3, style.fgcolor, 0, thickness);
-                break;
-            }
-
-            case BulletType::CheckBox: {
-                auto startpos = token.Bounds.start(initpos);
-                auto checkpos = ImVec2{ startpos.x + (bulletsz * 0.25f), startpos.y + (bulletsz * 0.25f) };
-                bulletsz *= 0.75f;
-                auto bsz3 = (bulletsz * 0.25f);
-                auto thickness = bulletsz * 0.25f;
-                ImVec2 points[3];
-                points[0] = { checkpos.x, checkpos.y + (2.5f * bsz3) };
-                points[1] = { checkpos.x + (bulletsz * 0.3333f), checkpos.y + bulletsz };
-                points[2] = { checkpos.x + bulletsz, checkpos.y + bsz3 };
-                drawList->AddPolyline(points, 3, style.fgcolor, 0, thickness);
-                drawList->AddRect(startpos, token.Bounds.end(initpos), style.fgcolor, thickness);
-                break;
-            }
-
-            case BulletType::Custom: {
                 if (config.DrawBullet != nullptr) {
-                    config.DrawBullet(token.Bounds.start(initpos), token.Bounds.end(initpos), style, token.ListItemIndex, 
+                    config.DrawBullet(token.Bounds.start(initpos), token.Bounds.end(initpos), style, token.ListItemIndex,
                         token.ListDepth, config.UserData);
                 }
-                else {
-                    ImVec2 center = token.Bounds.center(initpos);
-                    drawList->AddCircleFilled(center, bulletsz * 0.5f, style.fgcolor);
-                }
-                break;
+                else DrawBullet(drawList, initpos, token.Bounds, BulletType::FilledCircle, style.fgcolor, bulletsz);
             }
-
-            default:
-                break;
-            }
+            else DrawBullet(drawList, initpos, token.Bounds, style.list.itemStyle, style.fgcolor, bulletsz);
 
             DrawDebugRect(ContentTypeToken, drawList, token.Bounds.start(initpos), token.Bounds.end(initpos), config);
         }
@@ -1691,26 +1188,34 @@ namespace ImRichText
     }
 
     bool DrawSegment(ImDrawList* drawList, int lineidx, const SegmentData& segment,
-        ImVec2 initpos, ImVec2 bounds, const RenderConfig& config, 
+        ImVec2 initpos, ImVec2 bounds, const std::vector<StyleDescriptor>& styles, const RenderConfig& config,
         TooltipData& tooltip, AnimationData& animation, bool isMarquee)
     {
+        if (segment.Tokens.empty()) return true;
+
         auto popFont = false;
-        if (segment.Style.font.font != nullptr)
+        const auto& style = styles[segment.StyleIdx + 1];
+
+        if (style.font.font != nullptr)
         {
-            ImGui::PushFont(segment.Style.font.font);
+            ImGui::PushFont(style.font.font);
             popFont = true;
         }
 
-        const auto& style = segment.Style;
         auto drawTokens = true;
         auto startpos = segment.Bounds.start(initpos), endpos = segment.Bounds.end(initpos);
-        DrawBackground(drawList, startpos, endpos, style.gradient, style.bgcolor, config);
+
+        if ((style.blink && animation.isVisible) || !style.blink)
+            DrawBackground(drawList, startpos, endpos, style.gradient, style.bgcolor, config);
 
         for (const auto& token : segment.Tokens)
         {
-            if (drawTokens && !DrawToken(drawList, lineidx, token, initpos, bounds, style, 
+            if (drawTokens && !DrawToken(drawList, lineidx, token, initpos, bounds, style,
                 config, tooltip, animation, isMarquee))
-                drawTokens = false;
+            {
+                drawTokens = false; 
+                break;
+            }
         }
 
         DrawDebugRect(ContentTypeSegment, drawList, startpos, endpos, config);
@@ -1719,8 +1224,8 @@ namespace ImRichText
     }
 
     void DrawForegroundLayer(ImDrawList* drawList, ImVec2 initpos, ImVec2 bounds, 
-        const std::deque<DrawableLine>& lines, const RenderConfig& config, 
-        TooltipData& tooltip, AnimationData& animation)
+        const std::deque<DrawableLine>& lines, const std::vector<StyleDescriptor>& styles, 
+        const RenderConfig& config, TooltipData& tooltip, AnimationData& animation)
     {
         int listCountByDepth[IM_RICHTEXT_MAX_LISTDEPTH];
         int listDepth = -1;
@@ -1731,10 +1236,10 @@ namespace ImRichText
 
             for (const auto& segment : lines[lineidx].Segments)
             {
-                if (!DrawSegment(drawList, lineidx, segment, initpos, bounds, config, tooltip, animation, lines[lineidx].Marquee))
+                if (!DrawSegment(drawList, lineidx, segment, initpos, bounds, styles, config, tooltip, animation, lines[lineidx].Marquee))
                     break;
             }
-
+            
             DrawDebugRect(ContentTypeLine, drawList, lines[lineidx].Content.start(initpos), lines[lineidx].Content.end(initpos), config);
             if ((lines[lineidx].Content.top + lines[lineidx].height()) > (bounds.y + initpos.y)) break;
         }
@@ -1775,7 +1280,7 @@ namespace ImRichText
         
         if (animation.xoffsets.empty())
         {
-            animation.xoffsets.resize(drawables.ForegroundSegments.size());
+            animation.xoffsets.resize(drawables.ForegroundLines.size());
             std::fill(animation.xoffsets.begin(), animation.xoffsets.end(), 0.f);
         }
 
@@ -1785,14 +1290,14 @@ namespace ImRichText
         drawList->AddRectFilled(pos, endpos, config->DefaultBgColor);
 
         DrawBackgroundLayer(drawList, pos, bounds, drawables.BackgroundShapes, *config);
-        DrawForegroundLayer(drawList, pos, bounds, drawables.ForegroundSegments, *config, tooltip, animation);
+        DrawForegroundLayer(drawList, pos, bounds, drawables.ForegroundLines, drawables.StyleDescriptors, *config, tooltip, animation);
         DrawTooltip(drawList, tooltip, *config);
 
         if (!config->IsStrictHTML5 && (currFrameTime - animation.lastBlinkTime > IM_RICHTEXT_BLINK_ANIMATION_INTERVAL))
         {
             animation.isVisible = !animation.isVisible;
             animation.lastBlinkTime = currFrameTime;
-            if (config->NewFrameGenerated) config->NewFrameGenerated(config->UserData);
+            if (config->RequestFrame) config->RequestFrame(config->UserData);
         }
 
         if (currFrameTime - animation.lastMarqueeTime > IM_RICHTEXT_MARQUEE_ANIMATION_INTERVAL)
@@ -1800,12 +1305,12 @@ namespace ImRichText
             for (auto lineidx = 0; lineidx < (int)animation.xoffsets.size(); ++lineidx)
             {
                 animation.xoffsets[lineidx] += 1.f;
-                auto linewidth = drawables.ForegroundSegments[lineidx].width();
+                auto linewidth = drawables.ForegroundLines[lineidx].width();
 
                 if (animation.xoffsets[lineidx] >= linewidth)
                     animation.xoffsets[lineidx] = -linewidth;
 
-                if (config->NewFrameGenerated) config->NewFrameGenerated(config->UserData);
+                if (config->RequestFrame) config->RequestFrame(config->UserData);
             }
 
             animation.lastMarqueeTime = currFrameTime;
@@ -1832,19 +1337,22 @@ namespace ImRichText
 
     RenderConfig* GetDefaultConfig(ImVec2 Bounds, float defaultFontSize, float fontScale, bool skipDefaultFontLoading)
     {
+        DefaultRenderConfigInit = true;
+
         auto config = &DefaultRenderConfig;
         config->Bounds = Bounds;
-        config->GetFont = &GetFont;
         config->NamedColor = &GetColor;
-        config->GetTextSize = &GetTextSize;
         config->EscapeCodes.insert(config->EscapeCodes.end(), std::begin(EscapeCodes),
             std::end(EscapeCodes));
+
+#ifdef IM_RICHTEXT_DEFAULT_FONTS_AVAILABLE
+        config->GetFont = &GetFont;
+        config->GetTextSize = &GetTextSize;
         config->FontScale = fontScale;
         config->DefaultFontSize = defaultFontSize;
         config->MeterDefaultSize = { defaultFontSize * 5.0f, defaultFontSize };
-
-        DefaultRenderConfigInit = true;
         if (!skipDefaultFontLoading) LoadDefaultFonts(*config);
+#endif
         return config;
     }
 
@@ -1868,202 +1376,127 @@ namespace ImRichText
         if (it != RenderConfigs.end()) it->second.pop_back();
     }
 
-    enum ParseError
+    bool IsContentMultiline(TagType type)
     {
-        TagExtraction,
-        UnknownTag,
-        // ...
-    };
-
-    struct ITagVisitor
-    {
-        virtual bool TagStart(std::string_view) = 0;
-        virtual bool Attribute(std::string_view, std::optional<std::string_view>) = 0;
-        virtual bool TagStartDone() = 0;
-        virtual bool Content(std::string_view) = 0;
-        virtual bool PreFormattedContent(std::string_view) = 0;
-        virtual bool TagEnd(bool) = 0;
-        virtual void Finalize() = 0;
-
-        virtual void Error(std::string_view) = 0;
-        
-        virtual bool IsSelfTerminating(std::string_view) const = 0;
-        virtual bool IsPreformattedContent(std::string_view) const = 0;
-    };
-
-    void ParseRichText(const char* text, const char* textend, char TagStart, char TagEnd, ITagVisitor& visitor)
-    {
-        int end = (int)(textend - text), start = 0;
-        start = SkipSpace(text, start, end);
-        auto isPreformattedContent = false;
-        std::string_view lastTag = "";
-
-        for (auto idx = start; idx < end;)
+        switch (type)
         {
-            if (text[idx] == TagStart)
-            {
-                idx++;
-                auto tagStart = true, selfTerminatingTag = false;
-                auto [currTag, status] = ExtractTag(text, end, TagEnd, idx, tagStart);
-                if (!status) { visitor.Error(currTag); return; }
-
-                isPreformattedContent = visitor.IsPreformattedContent(currTag);
-                if (isPreformattedContent) lastTag = currTag;
-
-                if (tagStart)
-                {
-                    //LOG("Entering Tag: <%.*s>\n", (int)currTag.size(), currTag.data());
-                    if (!visitor.TagStart(currTag)) return;
-
-                    while ((idx < end) && (text[idx] != TagEnd) && (text[idx] != '/'))
-                    {
-                        auto begin = idx;
-                        while ((idx < end) && (text[idx] != '=') && !std::isspace(text[idx]) && (text[idx] != '/')) idx++;
-
-                        if (text[idx] != '/')
-                        {
-                            auto attribName = std::string_view{ text + begin, (std::size_t)(idx - begin) };
-                            //LOG("Attribute: %.*s\n", (int)attribName.size(), attribName.data());
-
-                            idx = SkipSpace(text, idx, end);
-                            if (text[idx] == '=') idx++;
-                            idx = SkipSpace(text, idx, end);
-                            auto attribValue = GetQuotedString(text, idx, end);
-                            if (!visitor.Attribute(attribName, attribValue)) return;
-                        }
-                    }
-
-                    if (text[idx] == TagEnd) idx++;
-                    if (text[idx] == '/' && ((idx + 1) < end) && text[idx + 1] == TagEnd) idx += 2;
-                }
-
-                selfTerminatingTag = (text[idx - 2] == '/' && text[idx - 1] == TagEnd) || visitor.IsSelfTerminating(currTag);
-
-                if (selfTerminatingTag || !tagStart) {
-                    if (!visitor.TagEnd(selfTerminatingTag)) return;
-                }
-                else if (!selfTerminatingTag && tagStart)
-                    if (!visitor.TagStartDone()) return;
-            }
-            else
-            {
-                auto begin = idx;
-
-                if (isPreformattedContent)
-                {
-                    static char EndTag[64] = { 0 };
-                    EndTag[0] = TagStart; EndTag[1] = '/';
-                    std::memcpy(EndTag + 2, lastTag.data(), lastTag.size());
-                    EndTag[2u + lastTag.size()] = TagEnd;
-                    EndTag[3u + lastTag.size()] = 0;
-
-                    while (((idx + (int)(lastTag.size() + 3u)) < end) && 
-                        AreSame(std::string_view{ text + idx, lastTag.size() + 3u }, EndTag)) idx++;
-                    std::string_view content{ text + begin, (std::size_t)(idx - begin) };
-
-                    if (!visitor.Content(content)) return;
-                }
-                else
-                {
-                    while ((idx < end) && (text[idx] != TagStart)) idx++;
-                    std::string_view content{ text + begin, (std::size_t)(idx - begin) };
-                    //LOG("Processing content [%.*s]\n", (int)content.size(), content.data());
-
-                    if (!visitor.Content(content)) return;
-                }
-            }
+            case ImRichText::TagType::Span: [[fallthrough]];
+            case ImRichText::TagType::Subscript: [[fallthrough]];
+            case ImRichText::TagType::Superscript: [[fallthrough]];
+            case ImRichText::TagType::Marquee: return false;
+            default: return true;
         }
-
-        visitor.Finalize();
     }
 
     struct DefaultTagVisitor final : public ITagVisitor
     {
-        Drawables result;
-        SegmentStyle initialStyle = CreateDefaultStyle(config);
-        
         std::string_view currTag;
         char TabSpaces[IM_RICHTEXT_MAXTABSTOP] = { 0 };
 
-        DrawableLine line;
+        DrawableLine currline;
+        StyleDescriptor currstyle;
+        int styleIdx = -1;
         int listDepth = -1, blockquoteDepth = -1;
         int subscriptLevel = 0, superscriptLevel = 0;
         bool currentListIsNumbered = false;
-        float maxWidth = config.Bounds.x;
+        float maxWidth = 0.f;
         TagType tagType = TagType::Unknown;
         
         const RenderConfig& config;
+        Drawables& result;
 
-        DefaultTagVisitor(const RenderConfig& cfg) : config{ cfg }
+        DefaultTagVisitor(const RenderConfig& cfg, Drawables& res) 
+            : config{ cfg }, result{ res }
         {
             CurrentStackPos = -1;
             std::memset(TabSpaces, ' ', Clamp(config.TabStop, 0, IM_RICHTEXT_MAXTABSTOP - 1));
             std::memset(ListItemCountByDepths, 0, IM_RICHTEXT_MAX_LISTDEPTH);
             TabSpaces[config.TabStop] = 0;
+            result.StyleDescriptors.emplace_back(CreateDefaultStyle(config));
+            maxWidth = config.Bounds.x;
         }
 
         void Finalize()
         {
-            MoveToNextLine(SegmentStyle{}, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, true);
-            maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+            MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, true);
+            maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
 
-            for (auto& line : result.ForegroundSegments)
+            for (auto& line : result.ForegroundLines)
             {
                 if (line.Marquee) line.Content.width = maxWidth;
             }
 
-            //for (auto depth = 0; depth < IM_RICHTEXT_MAXDEPTH; ++depth)
-        //{
-        //    for (const auto& bound : BlockquoteStack[depth].bounds)
-        //    {
-        //        if (config.BlockquoteBarWidth > 1.f && config.DefaultBgColor.Value != config.BlockquoteBar.Value)
-        //            result.BackgroundShapes.emplace_back(BackgroundShape{ bound.first, ImVec2{ config.BlockquoteBarWidth, bound.second.y },
-        //                config.BlockquoteBar });
+            for (auto depth = 0; depth < IM_RICHTEXT_MAXDEPTH; ++depth)
+            {
+                for (const auto& bound : BlockquoteStack[depth].bounds)
+                {
+                    if (config.BlockquoteBarWidth > 1.f && config.DefaultBgColor.Value != config.BlockquoteBar.Value)
+                        result.BackgroundShapes.emplace_back(BackgroundShape{ bound.first, ImVec2{ config.BlockquoteBarWidth, bound.second.y },
+                            config.BlockquoteBar });
 
-        //        if (config.DefaultBgColor.Value != config.BlockquoteBg.Value)
-        //            result.BackgroundShapes.emplace_back(BackgroundShape{ ImVec2{ bound.first.x + config.BlockquoteBarWidth, bound.first.y },
-        //                bound.second, config.BlockquoteBg });
-        //    }
+                    if (config.DefaultBgColor.Value != config.BlockquoteBg.Value)
+                        result.BackgroundShapes.emplace_back(BackgroundShape{ ImVec2{ bound.first.x + config.BlockquoteBarWidth, bound.first.y },
+                            bound.second, config.BlockquoteBg });
+                }
 
-        //    // Create multi-line background shapes for each depth and reset original specifications
-        //    for (const auto& bgdata : BgLineStartIndex[depth])
-        //    {
-        //        BackgroundShape background;
-        //        auto& firstSegment = result.ForegroundSegments[bgdata.start.first].Segments[bgdata.start.second];
-        //        const auto& lastSegment = result.ForegroundSegments[bgdata.end.first].Segments[bgdata.end.second];
-        //        background.Start = { firstSegment.Bounds.left, firstSegment.Bounds.top };
-        //        background.End = { lastSegment.Bounds.left + lastSegment.Bounds.width,
-        //            lastSegment.Bounds.top + result.ForegroundSegments[bgdata.end.first].height() };
-        //        background.Color = firstSegment.Style.bgcolor;
-        //        background.Gradient = firstSegment.Style.gradient;
-        //        firstSegment.Style.bgcolor = IM_COL32_BLACK_TRANS;
-        //        firstSegment.Style.gradient = ColorGradient{};
-        //        result.BackgroundShapes.push_back(background);
-        //    }
-        //}
+                // Create multi-line background shapes for each depth and reset original specifications
+                for (const auto& bgdata : BackgroundSpans[depth])
+                {
+                    BackgroundShape background;
+                    auto& firstSegment = result.ForegroundLines[bgdata.span.start.first].Segments[bgdata.span.start.second];
+                    const auto& lastSegment = result.ForegroundLines[bgdata.span.end.first].Segments[bgdata.span.end.second];
+                    auto& firstStyle = result.StyleDescriptors[bgdata.styleIdx + 1];
+
+                    background.Start = { firstSegment.Bounds.left, firstSegment.Bounds.top };
+                    background.End = { lastSegment.Bounds.left + lastSegment.Bounds.width,
+                        lastSegment.Bounds.top + result.ForegroundLines[bgdata.span.end.first].height() };
+                    background.Color = firstStyle.bgcolor;
+                    background.Gradient = firstStyle.gradient;
+                    firstStyle.bgcolor = IM_COL32_BLACK_TRANS;
+                    firstStyle.gradient = ColorGradient{};
+                    result.BackgroundShapes.push_back(background);
+                }
+            }
+        }
+
+        StyleDescriptor& Style(int stackpos)
+        {
+            return stackpos < 0 ? result.StyleDescriptors.front() : result.StyleDescriptors[TagStack[stackpos].styleIdx + 1];
         }
 
         bool TagStart(std::string_view tag)
         {
-            LOG("Entering Tag: <%.*s>\n", (int)currTag.size(), currTag.data());
-            tagType = GetTagType(currTag);
-            PushTag(currTag, initialStyle);
+            if (!IsContentMultiline(tagType) && AreSame(tag, "br")) return true;
+
+            LOG("Entering Tag: <%.*s>\n", (int)tag.size(), tag.data());
+            currTag = tag;
+            tagType = GetTagType(tag);
+            
+            PushTag(currTag, styleIdx);
             if (tagType == TagType::Superscript) superscriptLevel++;
             else if (tagType == TagType::Subscript) subscriptLevel++;
+
+            if (CurrentStackPos >= 0 && TagStack[CurrentStackPos].tag != currTag)
+                ERROR("Tag mismatch...");
+            return true;
         }
         
         bool Attribute(std::string_view name, std::optional<std::string_view> value)
         {
-            auto& currentStyle = TagStack[CurrentStackPos].second;
-            const auto& parentStyle = CurrentStackPos == 0 ? initialStyle : TagStack[CurrentStackPos - 1].second;
-
-            currentStyle.propsSpecified = RecordTagProperties(tagType, name, value, currentStyle, parentStyle, config);
-            SetImplicitStyleProps(tagType, currTag, currentStyle, parentStyle, line, currentStyle.propsSpecified, config);
+            LOG("Reading attribute: %.*s\n", (int)name.size(), name.data());
+            const auto& parentStyle = Style(CurrentStackPos - 1);
+            currstyle.propsSpecified = RecordTagProperties(tagType, name, value, currstyle, parentStyle, config);
+            return true;
         }
 
         bool TagStartDone()
         {
-            auto& currentStyle = TagStack[CurrentStackPos].second;
+            const auto& parentStyle = Style(CurrentStackPos - 1);
+            SetImplicitStyleProps(tagType, currTag, currstyle, parentStyle, currline, currstyle.propsSpecified, config);
+            if (IsStyleSupported(tagType) && currstyle.propsSpecified != 0) 
+                result.StyleDescriptors.emplace_back(currstyle);
+            styleIdx = ((int)result.StyleDescriptors.size() - 2);
+            auto& currentStyle = Style(CurrentStackPos);
 
             if (tagType == TagType::List)
             {
@@ -2074,11 +1507,11 @@ namespace ImRichText
                 tagType == TagType::RawText || tagType == TagType::ListItem ||
                 tagType == TagType::CodeBlock || tagType == TagType::Marquee)
             {
-                line = MoveToNextLine(currentStyle, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, true);
-                maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, true);
+                maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
 
                 if (tagType == TagType::Paragraph && config.ParagraphStop > 0 && config.GetTextSize)
-                    line.Offset.left += config.GetTextSize(std::string_view{ LineSpaces,
+                    currline.Offset.left += config.GetTextSize(std::string_view{ LineSpaces,
                         (std::size_t)std::min(config.ParagraphStop, IM_RICHTEXT_MAXTABSTOP) }, currentStyle.font.font).x;
                 else if (tagType == TagType::ListItem)
                 {
@@ -2089,81 +1522,86 @@ namespace ImRichText
                         TokenType::ListItemNumbered;
                     token.ListDepth = listDepth;
                     token.ListItemIndex = ListItemCountByDepths[listDepth];
-                    AddToken(line, token, currentStyle.propsSpecified, config);
-                    AddSegment(line, currentStyle, config);
+                    AddToken(currline, token, currentStyle.propsSpecified, result.StyleDescriptors, config);
+                    AddSegment(currline, styleIdx, result.StyleDescriptors, config);
                 }
             }
             else if (tagType == TagType::Blockquote)
             {
                 blockquoteDepth++;
-                line = MoveToNextLine(currentStyle, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, true);
-                maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, true);
+                maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
                 auto& start = BlockquoteStack[blockquoteDepth].bounds.emplace_back();
-                start.first = ImVec2{ line.Content.left, line.Content.top };
+                start.first = ImVec2{ currline.Content.left, currline.Content.top };
             }
             else if (tagType == TagType::Quotation)
             {
                 Token token;
                 token.Type = TokenType::Text;
                 token.Content = "\"";
-                AddToken(line, token, currentStyle.propsSpecified, config);
+                AddToken(currline, token, currentStyle.propsSpecified, result.StyleDescriptors, config);
             }
             else if (tagType == TagType::Meter)
             {
                 Token token;
                 token.Type = TokenType::Meter;
-                AddToken(line, token, currentStyle.propsSpecified, config);
+                AddToken(currline, token, currentStyle.propsSpecified, result.StyleDescriptors, config);
             }
 
-            if (line.Segments.empty() || tagType == TagType::Blink)
-                AddSegment(line, currentStyle, config);
+            if (currline.Segments.empty() || tagType == TagType::Blink)
+                AddSegment(currline, styleIdx, result.StyleDescriptors, config);
 
-            auto& segment = line.Segments.back();
+            auto& segment = currline.Segments.back();
             segment.SubscriptDepth = subscriptLevel;
             segment.SuperscriptDepth = superscriptLevel;
 
-            if ((currentStyle.propsSpecified & StyleBackground) != 0)
+            if ((currentStyle.propsSpecified & StyleBackground) != 0 &&
+                tagType == TagType::Paragraph)
             {
                 // The current line is `line` which is not yet added, hence record .size()
-                auto& bgidx = BgLineStartIndex[CurrentStackPos].emplace_back();
-                bgidx.start.first = (int)result.ForegroundSegments.size();
-                bgidx.start.second = (int)line.Segments.size() - 1;
+                auto& bgidx = BackgroundSpans[CurrentStackPos].emplace_back();
+                bgidx.span.start.first = (int)result.ForegroundLines.size();
+                bgidx.span.start.second = (int)currline.Segments.size() - 1;
+                bgidx.styleIdx = styleIdx;
             }
+
+            return true;
         }
 
         bool Content(std::string_view content)
         {
             // Ignore newlines, tabs & consecutive spaces
             auto curridx = 0, start = 0;
-            auto& currentStyle = TagStack[CurrentStackPos].second;
+            auto& currentStyle = Style(CurrentStackPos);
+            LOG("Processing content [%.*s]\n", (int)content.size(), content.data());
 
             while (curridx < (int)content.size())
             {
                 if (content[curridx] == '\n')
                 {
-                    GenerateTextToken(line, content, curridx, start, config);
+                    GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
                     while (curridx < (int)content.size() && content[curridx] == '\n') curridx++;
                     start = curridx;
                 }
                 else if (content[curridx] == '\t')
                 {
-                    GenerateTextToken(line, content, curridx, start, config);
+                    GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
                     while (curridx < (int)content.size() && content[curridx] == '\t') curridx++;
                     start = curridx;
                 }
                 else if (content[curridx] == config.EscapeSeqStart)
                 {
-                    GenerateTextToken(line, content, curridx, start, config);
+                    GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
 
                     curridx++;
                     auto [hasEscape, isNewLine] = AddEscapeSequences(content, curridx, config.EscapeCodes,
-                        config.EscapeSeqStart, config.EscapeSeqEnd, line, start);
+                        config.EscapeSeqStart, config.EscapeSeqEnd, currline, start);
                     curridx = start;
 
                     if (isNewLine && !subscriptLevel && !superscriptLevel)
                     {
-                        line = MoveToNextLine(currentStyle, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, true);
-                        maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                        currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, true);
+                        maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
                     }
 
                     if (hasEscape) continue;
@@ -2172,7 +1610,7 @@ namespace ImRichText
                 {
                     curridx++;
 
-                    if (curridx - start > 0) GenerateTextToken(line, content, curridx, start, config);
+                    if (curridx - start > 0) GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
                     curridx = SkipSpace(content, curridx);
                     start = curridx;
                     continue;
@@ -2181,13 +1619,14 @@ namespace ImRichText
                 curridx++;
             }
 
-            if (curridx > start) GenerateTextToken(line, content, curridx, start, config);
+            if (curridx > start) GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
+            return true;
         }
 
         bool PreFormattedContent(std::string_view content)
         {
             auto curridx = 0, start = 0;
-            auto& currentStyle = TagStack[CurrentStackPos].second;
+            auto& currentStyle = Style(CurrentStackPos);
 
             while (curridx < (int)content.size())
             {
@@ -2195,14 +1634,14 @@ namespace ImRichText
                 {
                     if (!subscriptLevel && !superscriptLevel)
                     {
-                        GenerateTextToken(line, content, curridx, start, config);
-                        line = MoveToNextLine(currentStyle, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, true);
-                        maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                        GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
+                        currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, true);
+                        maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
                         start = curridx;
                     }
                     else
                     {
-                        GenerateTextToken(line, content, curridx, start, config);
+                        GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
                         start = curridx;
                     }
                 }
@@ -2210,19 +1649,21 @@ namespace ImRichText
                 ++curridx;
             }
 
-            if (curridx > start)
-            {
-                GenerateTextToken(line, content, curridx, start, config);
-            }
+            if (curridx > start) GenerateTextToken(currline, content, curridx, start, result.StyleDescriptors, config);
+            return true;
         }
 
-        bool TagEnd(bool selfTerminatingTag)
+        bool TagEnd(std::string_view tag, bool selfTerminatingTag)
         {
+            if (!IsContentMultiline(tagType) && AreSame(tag, "br")) return true;
+
             // pop stye properties and reset
-            auto hasBgSpecified = (TagStack[CurrentStackPos].second.propsSpecified & StyleBackground) != 0;
+            auto hasBgSpecified = (Style(CurrentStackPos).propsSpecified & StyleBackground) != 0;
             PopTag(!selfTerminatingTag);
-            const auto& styleprops = CurrentStackPos >= 0 ?
-                TagStack[CurrentStackPos].second : initialStyle;
+
+            const auto& styleprops = Style(CurrentStackPos);
+            styleIdx = CurrentStackPos >= 0 ? TagStack[CurrentStackPos].styleIdx : -1;
+            auto segmentAdded = false;
             LOG("Exiting Tag: <%.*s>\n", (int)currTag.size(), currTag.data());
 
             if (tagType == TagType::List || tagType == TagType::Paragraph || tagType == TagType::Header ||
@@ -2235,75 +1676,80 @@ namespace ImRichText
                     listDepth--;
                 }
 
-                line = MoveToNextLine(styleprops, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, false);
-                maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, false);
+                maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
 
                 if (tagType == TagType::Blockquote)
                 {
                     assert(!BlockquoteStack[blockquoteDepth].bounds.empty());
                     auto& bounds = BlockquoteStack[blockquoteDepth].bounds.back();
-                    const auto& lastLine = result.ForegroundSegments[result.ForegroundSegments.size() - 2u];
+                    const auto& lastLine = result.ForegroundLines[result.ForegroundLines.size() - 2u];
                     bounds.second = ImVec2{ lastLine.width() + bounds.first.x, lastLine.Content.top + lastLine.height() };
                     blockquoteDepth--;
+
+                    AddSegment(currline, styleIdx, result.StyleDescriptors, config);
+                    segmentAdded = true;
                 }
                 else if (tagType == TagType::Header)
                 {
                     // Add properties for horizontal line below header
-                    auto& style = line.Segments.back().Style;
+                    auto& style = result.StyleDescriptors[currline.Segments.back().StyleIdx + 1];
                     style.height = 1.f;
                     style.fgcolor = config.HeaderLineColor;
                     style.padding.top = style.padding.bottom = config.HrVerticalMargins;
 
                     Token token;
                     token.Type = TokenType::HorizontalRule;
-                    AddToken(line, token, NoStyleChange, config);
+                    AddToken(currline, token, NoStyleChange, result.StyleDescriptors, config);
 
                     // Move to next line for other content
-                    line = MoveToNextLine(styleprops, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, false);
-                    maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                    currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, false);
+                    maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
                 }
             }
             else if (tagType == TagType::Hr)
             {
                 // Since hr style is popped, refer to next item in stack
-                auto& prevstyle = TagStack[CurrentStackPos + 1].second;
+                auto& prevstyle = Style(CurrentStackPos + 1);
                 prevstyle.padding.top = prevstyle.padding.bottom = config.HrVerticalMargins;
-                line = MoveToNextLine(prevstyle, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, false);
-                maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, false);
+                maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
 
                 Token token;
                 token.Type = TokenType::HorizontalRule;
-                AddToken(line, token, NoStyleChange, config);
+                AddToken(currline, token, NoStyleChange, result.StyleDescriptors, config);
 
-                line = MoveToNextLine(styleprops, tagType, listDepth, blockquoteDepth, line, result.ForegroundSegments, config, true);
-                maxWidth = std::max(maxWidth, result.ForegroundSegments.back().Content.width);
+                currline = MoveToNextLine(tagType, listDepth, blockquoteDepth, styleIdx, currline, result.ForegroundLines, result.StyleDescriptors, config, true);
+                maxWidth = std::max(maxWidth, result.ForegroundLines.back().Content.width);
             }
             else if (tagType == TagType::Quotation)
             {
                 Token token;
                 token.Type = TokenType::Text;
                 token.Content = "\"";
-                AddToken(line, token, NoStyleChange, config);
+                AddToken(currline, token, NoStyleChange, result.StyleDescriptors, config);
             }
             else
             {
-                AddSegment(line, styleprops, config);
+                AddSegment(currline, styleIdx, result.StyleDescriptors, config);
+                segmentAdded = true;
 
                 if (tagType == TagType::Superscript) superscriptLevel--;
                 else if (tagType == TagType::Subscript) subscriptLevel--;
             }
 
             // If background spans multiple lines, record it separately
-            if (!selfTerminatingTag && !BgLineStartIndex[CurrentStackPos + 1].empty() && hasBgSpecified)
+            if (!selfTerminatingTag && !BackgroundSpans[CurrentStackPos + 1].empty() && hasBgSpecified)
             {
-                auto& lastLineIdx = BgLineStartIndex[CurrentStackPos + 1].back();
-                lastLineIdx.end.first = (int)result.ForegroundSegments.size();
-                lastLineIdx.end.second = (int)line.Segments.size() - 1;
+                auto& lastLineIdx = BackgroundSpans[CurrentStackPos + 1].back();
+                lastLineIdx.span.end.first = (int)result.ForegroundLines.size();
+                lastLineIdx.span.end.second = std::max(0, (int)currline.Segments.size() - (segmentAdded ? 2 : 1));
             }
 
-            if (selfTerminatingTag) TagStack[CurrentStackPos + 1] = std::make_pair(std::string_view{}, SegmentStyle{});
-            currTag = CurrentStackPos == -1 ? "" : TagStack[CurrentStackPos].first;
+            if (selfTerminatingTag) TagStack[CurrentStackPos + 1] = StackData{};
+            currTag = CurrentStackPos == -1 ? "" : TagStack[CurrentStackPos].tag;
             tagType = TagType::Unknown;
+            return true;
         }
 
         void Error(std::string_view tag)
@@ -2325,7 +1771,7 @@ namespace ImRichText
     Drawables GetDrawables(const char* text, const char* textend, const RenderConfig& config)
     {
         Drawables result;
-        DefaultTagVisitor visitor{ config };
+        DefaultTagVisitor visitor{ config, result };
         ParseRichText(text, textend, config.TagStart, config.TagEnd, visitor);
         return result;
     }
@@ -2338,7 +1784,7 @@ namespace ImRichText
         if (bounds.x < 0.f)
         {
             float width = 0.f;
-            for (const auto& line : drawables.ForegroundSegments)
+            for (const auto& line : drawables.ForegroundLines)
                 width = std::max(width, line.width() + line.Content.left);
             for (const auto& bg : drawables.BackgroundShapes)
                 width = std::max(width, bg.End.x);
@@ -2349,9 +1795,9 @@ namespace ImRichText
         {
             auto fgheight = 0.f, bgheight = 0.f;
 
-            if (!drawables.ForegroundSegments.empty())
+            if (!drawables.ForegroundLines.empty())
             {
-                const auto& lastFg = drawables.ForegroundSegments.back();
+                const auto& lastFg = drawables.ForegroundLines.back();
                 fgheight = lastFg.height() + lastFg.Content.top;
             }
