--- imrichtextcolor.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "imgui.h"
-
-namespace ImRichText
-{
-	[[nodiscard]] ImColor GetColor(const char* name, void*);
-}
